+++ src/LibPARI.jl
@@ -79,7 +79,7 @@ if VERSION >= v"1.11"
     eval(
         Meta.parse(
             "public PARI, PariObject, PariType, PariErr, PariConvertible, " *
-            "ConversionError, gen_convert, gentype, gen_from, " *
+            "ConversionError, GenArg, gen_convert, gentype, gen_from, " *
             "protected_call, isexact, nbits2prec, default_precision, " *
             "isprime, nextprime, prevprime, factor, factors, " *
             "is_initialized, library_state, stack_size, serve_mcp",